+++ mautrix_googlechat/config.py
@@ -55,6 +55,8 @@ class Config(BaseBridgeConfig):
         copy("bridge.encryption.key_sharing.require_cross_signing")
         copy("bridge.encryption.key_sharing.require_verification")
         copy("bridge.delivery_receipts")
+        copy("bridge.delivery_error_reports")
+        copy("bridge.message_status_events")
         copy("bridge.federate_rooms")
         copy("bridge.backfill.invite_own_puppet")
         copy("bridge.backfill.initial_thread_limit")